src/test/java/org.example/WebFilterTest.java
package org.example;

import java.io.File;
import java.io.FileInputStream;
import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.HashMap;
import java.util.HashSet;
import java.util.List;
import java.util.Map;
import java.util.Set;
import java.util.TreeMap;
import java.util.function.Predicate;
import java.util.stream.Collectors;

import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.ServletRequest;
import javax.servlet.ServletResponse;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletRequestWrapper;

import org.example.WebFilter;


import org.junit.jupiter.api.Test;
import org.mockito.ArgumentCaptor;
import org.mockito.Mockito;
import org.springframework.boot.actuate.web.mappings.MappingsEndpoint;
import org.springframework.boot.actuate.web.mappings.servlet.DispatcherServletMappingDescription;
import org.springframework.boot.actuate.web.mappings.servlet.DispatcherServletMappingDetails;
import org.springframework.boot.actuate.web.mappings.servlet.RequestMappingConditionsDescription;

import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.junit.jupiter.api.Assertions.fail;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.never;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;



class WebFilterTest {

    @Test
    public void test_castServletRequestToHttpServletRequest() throws ServletException, IOException {
        ServletRequest servletRequest = mock(ServletRequest.class);
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        HttpServletRequest req = mock(HttpServletRequest.class);
        when(servletRequest instanceof HttpServletRequest).thenReturn(true);
        when((HttpServletRequest) servletRequest).thenReturn(req);

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        try {
            webFilter.doFilter(servletRequest, servletResponse, filterChain);
        } catch (IOException | ServletException e) {
            fail("Exception thrown");
        }

        verify(filterChain).doFilter(req, servletResponse);
    }

    // The method retrieves the request URI from the HttpServletRequest.
    @Test
    public void test_retrieveRequestURI() {
        HttpServletRequest req = mock(HttpServletRequest.class);
        when(req.getRequestURI()).thenReturn("/example");

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        try {
            webFilter.doFilter(req, mock(ServletResponse.class), mock(FilterChain.class));
        } catch (IOException | ServletException e) {
            fail("Exception thrown");
        }

        verify(req).getRequestURI();
    }

    // The ServletRequest is null.
    @Test
    public void test_ServletRequestIsNull() {
        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        try {
            webFilter.doFilter(null, mock(ServletResponse.class), mock(FilterChain.class));
        } catch (IOException | ServletException e) {
            fail("Exception thrown");
        }

        // No assertion needed, as the method should not throw an exception
    }

    // The ServletResponse is null.
    @Test
    public void test_ServletResponseIsNull() {
        HttpServletRequest req = mock(HttpServletRequest.class);

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        try {
            webFilter.doFilter(req, null, mock(FilterChain.class));
        } catch (IOException | ServletException e) {
            fail("Exception thrown");
        }

        // No assertion needed, as the method should not throw an exception
    }

    // The FilterChain is null.
    @Test
    public void test_FilterChainIsNull() {
        HttpServletRequest req = mock(HttpServletRequest.class);

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        try {
            webFilter.doFilter(req, mock(ServletResponse.class), null);
        } catch (IOException | ServletException e) {
            fail("Exception thrown");
        }

        // No assertion needed, as the method should not throw an exception
    }

    // The method creates a new HttpServletRequestWrapper if the request URI does not start with any of the requested paths.
    @Test
    public void test_createHttpServletRequestWrapper() {
        // Create a mock MappingsEndpoint
        MappingsEndpoint mappingsEndpoint = mock(MappingsEndpoint.class);

        // Create a mock HttpServletRequest
        HttpServletRequest servletRequest = mock(HttpServletRequest.class);
        when(servletRequest.getRequestURI()).thenReturn("/example");

        // Create a mock ServletResponse
        ServletResponse servletResponse = mock(ServletResponse.class);

        // Create a mock FilterChain
        FilterChain filterChain = mock(FilterChain.class);

        // Create an instance of WebFilter
        WebFilter webFilter = new WebFilter(mappingsEndpoint);

        try {
            // Call the doFilter method
            webFilter.doFilter(servletRequest, servletResponse, filterChain);

            // Verify that a new HttpServletRequestWrapper is created
            ArgumentCaptor<HttpServletRequest> requestCaptor = ArgumentCaptor.forClass(HttpServletRequest.class);
            verify(filterChain).doFilter(requestCaptor.capture(), eq(servletResponse));

            HttpServletRequest capturedRequest = requestCaptor.getValue();
            assertTrue(capturedRequest instanceof HttpServletRequestWrapper);

            // Verify that the getRequestURI method of the HttpServletRequestWrapper returns the modified URI
            assertEquals("/config/example", capturedRequest.getRequestURI());
        } catch (IOException | ServletException e) {
            fail("An exception occurred: " + e.getMessage());
        }
    }

    // The method correctly sets the request URI in the new HttpServletRequestWrapper.
    @Test
    public void test_correctly_sets_request_uri() {
        HttpServletRequest servletRequest = mock(HttpServletRequest.class);
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        when(servletRequest.getRequestURI()).thenReturn("/example");

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        try {
            webFilter.doFilter(servletRequest, servletResponse, filterChain);
        } catch (IOException | ServletException e) {
            fail("Exception thrown: " + e.getMessage());
        }

        ArgumentCaptor<HttpServletRequest> requestCaptor = ArgumentCaptor.forClass(HttpServletRequest.class);
        try {
            verify(filterChain).doFilter(requestCaptor.capture(), eq(servletResponse));
        } catch (IOException | ServletException e) {
            fail("Exception thrown: " + e.getMessage());
        }

        HttpServletRequest capturedRequest = requestCaptor.getValue();
        assertEquals("/config/example", capturedRequest.getRequestURI());
    }

    // The HttpServletRequest is null.
    @Test
    public void test_nullHttpServletRequest() throws IOException, ServletException {
        // Arrange
        ServletRequest servletRequest = null;
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));

        // Act
        assertThrows(ClassCastException.class, () -> {
            webFilter.doFilter(servletRequest, servletResponse, filterChain);
        });

        // Assert
        verify(filterChain, never()).doFilter(any(), any());
    }

    // The mappingsEndpoint.mappings().getContexts().values() returns null.
    @Test
    public void test_mappingsEndpointReturnsNull1() {
        MappingsEndpoint mappingsEndpoint = mock(MappingsEndpoint.class);
        when(mappingsEndpoint.mappings()).thenReturn(null);

        WebFilter webFilter = new WebFilter(mappingsEndpoint);

        HttpServletRequest servletRequest = mock(HttpServletRequest.class);
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        assertThrows(NullPointerException.class, () -> {
            webFilter.doFilter(servletRequest, servletResponse, filterChain);
        });
    }

    // The mappingsEndpoint.mappings() returns null.
    @Test
    public void test_mappingsEndpointReturnsNull2() throws ServletException, IOException {
        MappingsEndpoint mappingsEndpoint = mock(MappingsEndpoint.class);
        when(mappingsEndpoint.mappings()).thenReturn(null);

        WebFilter webFilter = new WebFilter(mappingsEndpoint);

        HttpServletRequest servletRequest = mock(HttpServletRequest.class);
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        assertDoesNotThrow(() -> webFilter.doFilter(servletRequest, servletResponse, filterChain));

        verify(filterChain).doFilter(servletRequest, servletResponse);
    }

    // The mappingsEndpoint.mappings().getContexts().values().stream() returns null.
    @Test
    public void test_mappingsEndpointReturnsNull() throws ServletException, IOException {
        // Create a mock MappingsEndpoint
        MappingsEndpoint mappingsEndpointMock = mock(MappingsEndpoint.class);

        // Set up the mock to return null for mappings().getContexts().values().stream()
        when(mappingsEndpointMock.mappings().getContexts().values().stream()).thenReturn(null);

        // Create a WebFilter instance with the mock MappingsEndpoint
        WebFilter webFilter = new WebFilter(mappingsEndpointMock);

        // Create mock ServletRequest, ServletResponse, and FilterChain
        ServletRequest servletRequestMock = mock(ServletRequest.class);
        ServletResponse servletResponseMock = mock(ServletResponse.class);
        FilterChain filterChainMock = mock(FilterChain.class);

        // Call the doFilter method
        try {
            webFilter.doFilter(servletRequestMock, servletResponseMock, filterChainMock);
        } catch (IOException | ServletException e) {
            // Handle exception if necessary
        }

        // Assert that the getRequestURI method was not called on the HttpServletRequest
        verify((HttpServletRequest) servletRequestMock, never()).getRequestURI();

        // Assert that the doFilter method was called with the modified HttpServletRequest
        verify(filterChainMock).doFilter(any(HttpServletRequest.class), eq(servletResponseMock));
    }



    // The mappingsEndpoint is null.
    @Test
    public void test_mappingsEndpointIsNull() {
        MappingsEndpoint mappingsEndpoint = null;
        WebFilter webFilter = new WebFilter(mappingsEndpoint);

        HttpServletRequest servletRequest = mock(HttpServletRequest.class);
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        assertThrows(NullPointerException.class, () -> {
            webFilter.doFilter(servletRequest, servletResponse, filterChain);
        });
    }


    // The request URI is null.
    @Test
    public void test_requestURIIsNull() {
        HttpServletRequest servletRequest = mock(HttpServletRequest.class);
        ServletResponse servletResponse = mock(ServletResponse.class);
        FilterChain filterChain = mock(FilterChain.class);

        when(servletRequest.getRequestURI()).thenReturn(null);

        WebFilter webFilter = new WebFilter(mock(MappingsEndpoint.class));
        assertThrows(NullPointerException.class, () -> {
            webFilter.doFilter(servletRequest, servletResponse, filterChain);
        });
    }

    // The requested paths are null.
    @Test
    public void test_requested_paths_null() throws ServletException, IOException {
        // Create a mock MappingsEndpoint
        MappingsEndpoint mappingsEndpoint = mock(MappingsEndpoint.class);

        // Set up the mock to return null for mappings().getContexts()
        when(mappingsEndpoint.mappings().getContexts()).thenReturn(null);

        // Create a mock ServletRequest
        ServletRequest servletRequest = mock(ServletRequest.class);

        // Create a mock ServletResponse
        ServletResponse servletResponse = mock(ServletResponse.class);

        // Create a mock FilterChain
        FilterChain filterChain = mock(FilterChain.class);

        // Create a WebFilter instance with the mock MappingsEndpoint
        WebFilter webFilter = new WebFilter(mappingsEndpoint);

        // Call the doFilter method with the mock objects
        webFilter.doFilter(servletRequest, servletResponse, filterChain);

        // Verify that the filterChain.doFilter method was called with the modified request
        verify(filterChain).doFilter(any(HttpServletRequestWrapper.class), eq(servletResponse));
    }



    void finder(){


    }

}